+++ brightwind/load/load.py
@@ -447,6 +447,8 @@ def _append_files_together(source_folder, assembled_file_name, file_type, append
 
 class LoadBrightdata:
     """
+    ** BRIGHTDATA IS NO LONGER SUPPORTED WITH NEW DATA. THIS CLASS WILL BE REMOVED IN V3.0. **
+
     LoadBrightdata allows you to pull meta data and timeseries data of reanalysis datasets from brightdata. This
     is a fast way to get access to the available reanalysis datasets.
 
@@ -466,6 +468,7 @@ class LoadBrightdata:
 
     _BASE_URI = 'http://api.brightwindanalysis.com/brightdata/'
     # _BASE_URI = 'http://localhost:5000/'
+    _SHOW_DEPRECATION_WARNING = True
 
     class Node:
         """
@@ -497,6 +500,15 @@ class LoadBrightdata:
         :param query_params: dictionary of the query parameters to be sent
         :return: List(Node)
         """
+
+        warnings.warn(
+            "BrightDATA was deprecated in v2.4.0 and will be removed in v3.0.0. "
+            "Support for reanalysis data has moved to BrightHub. "
+            "Please use LoadBrightHub instead of LoadBrightdata to continue accessing reanalysis data.",
+            DeprecationWarning,
+            stacklevel=3
+        )
+
         username = utils.get_environment_variable('BRIGHTDATA_USERNAME')
         password = utils.get_environment_variable('BRIGHTDATA_PASSWORD')
 
@@ -548,6 +560,12 @@ class LoadBrightdata:
     @staticmethod
     def timeseries(dataset, lat, long, nearest=None, from_date=None, to_date=None, variables=None):
         """
+            ** BRIGHTDATA IS NO LONGER SUPPORTED WITH NEW DATA. THIS FUNCTION IS DEPRECIATED IN V2.4.0 AND WILL BE
+            REMOVED IN V3.0.
+
+            SUPPORT FOR REANALYSIS DATA HAS MOVED TO BRIGHTHUB. PLEASE USE LOADBRIGHTHUB INSTEAD OF LOADBRIGHTDATA TO
+            CONTINUE ACCESSING REANALYSIS DATA. **
+
             Retrieve timeseries datasets available from brightdata. Returns a list of Node objects in order
             of closest distance to the requested lat, long.
 
@@ -632,6 +650,12 @@ class LoadBrightdata:
     @staticmethod
     def monthly_means(dataset, lat, long, nearest=None, from_date=None, to_date=None, variables=None):
         """
+            ** BRIGHTDATA IS NO LONGER SUPPORTED WITH NEW DATA. THIS FUNCTION IS DEPRECIATED IN V2.4.0 AND WILL BE
+            REMOVED IN V3.0.
+
+            SUPPORT FOR REANALYSIS DATA HAS MOVED TO BRIGHTHUB. PLEASE USE LOADBRIGHTHUB INSTEAD OF LOADBRIGHTDATA TO
+            CONTINUE ACCESSING REANALYSIS DATA. **
+
             Retrieve monthly means from brightdata datasets such as merra2 and era5. Returns a list of Node objects
             in order of closest distance to the requested lat, long. Monthly coverage is also returned in the data.
 
@@ -713,6 +737,12 @@ class LoadBrightdata:
     @staticmethod
     def momm(dataset, lat, long, nearest=None, from_date=None, to_date=None, variables=None):
         """
+            ** BRIGHTDATA IS NO LONGER SUPPORTED WITH NEW DATA. THIS FUNCTION IS DEPRECIATED IN V2.4.0 AND WILL BE
+            REMOVED IN V3.0.
+
+            SUPPORT FOR REANALYSIS DATA HAS MOVED TO BRIGHTHUB. PLEASE USE LOADBRIGHTHUB INSTEAD OF LOADBRIGHTDATA TO
+            CONTINUE ACCESSING REANALYSIS DATA. **
+
             Retrieve the mean of monthly means from brightdata datasets such as merra2 and era5. Returns a list of
             Node objects in order of closest distance to the requested lat, long.
 
@@ -796,6 +826,12 @@ class LoadBrightdata:
     def monthly_norms(dataset, lat, long, nearest=None, from_date=None, to_date=None, ref_from_date=None,
                       ref_to_date=None, ref_no_years=None, variables=None):
         """
+        ** BRIGHTDATA IS NO LONGER SUPPORTED WITH NEW DATA. THIS FUNCTION IS DEPRECIATED IN V2.4.0 AND WILL BE
+        REMOVED IN V3.0.
+
+        SUPPORT FOR REANALYSIS DATA HAS MOVED TO BRIGHTHUB. PLEASE USE LOADBRIGHTHUB INSTEAD OF LOADBRIGHTDATA TO
+        CONTINUE ACCESSING REANALYSIS DATA. **
+
         Return the monthly mean wind speeds normalised to a specific reference period. The reference period can be
         between two specific dates or it could be a number of rolling years preceding each month of interest.