+++ CHANGELOG.md
@@ -15,7 +15,6 @@ Additional labels for pre-release and build metadata are available as extensions
 1. Fixed pandas<3.0.0 and numpy<2.3.1 dependencies, ([#458](https://github.com/brightwind-dev/brightwind/issues/458))
 1. Fixed pandas deprecating warnings that were linked to frequency strings, .groupby() and .map(). ([#407](https://github.com/brightwind-dev/brightwind/issues/407), [#415](https://github.com/brightwind-dev/brightwind/issues/415) and [#445](https://github.com/brightwind-dev/brightwind/issues/445))
 
-
 ### Bug Fixes
 1. 
 
@@ -33,6 +32,9 @@ Additional labels for pre-release and build metadata are available as extensions
    1. Added deprecation warning the username and password method of authenticating in BrightHub will be removed in a future version.
 1. Updated `time_continuity_gaps()` in order to take an argument `minimum_gap_length` which allows the user to filter the time gaps returned. (Issue [#545](https://github.com/brightwind-dev/brightwind/issues/545))
 1. Updated `offset_timestamps()`to include last timestamp when date_to is unspecified, so that offset is applied to the entire record if date_to not specified. ([#504](https://github.com/brightwind-dev/brightwind/issues/504))
+
+### Deprecated
+1. `LoadBrightdata()` is deprecated and will be removed in version 3.0. Use `LoadBrightHub()` instead to continue accessing reanalysis data.
 
 
 ## [2.3.0]